Remove unnecessary casts from AppendBodyTo implementations
Agent-Logs-Url: https://github.com/0blu/vmangos-core/sessions/452d684d-4eb7-489f-a3b0-0cfef237f1cc

+++ src/game/Server/Packets/AuctionHouse.cpp
@@ -64,8 +64,8 @@ void WorldPackets::AuctionHouse::AuctionListItems::ReadFromWorldPacket(WorldPack
 
 void WorldPackets::AuctionHouse::AuctionRemovedNotification::AppendBodyTo(ByteBuffer& buffer) const
 {
-    buffer << uint32(auctionId);
-    buffer << uint32(itemTemplate);
-    buffer << uint32(randomPropertyId);
+    buffer << auctionId;
+    buffer << itemTemplate;
+    buffer << randomPropertyId;
 }
