+++ src/game/Server/Packets/GmTicket.cpp
@@ -43,20 +43,20 @@ void WorldPackets::GmTicket::GMSurveySubmit::ReadFromWorldPacket(WorldPacket& re
 
 void WorldPackets::GmTicket::GmTicketUpdateTextResponse::AppendBodyTo(ByteBuffer& buffer) const
 {
-    buffer << uint32(response);
+    buffer << response;
 }
 
 void WorldPackets::GmTicket::GmTicketDeleteTicketResponse::AppendBodyTo(ByteBuffer& buffer) const
 {
-    buffer << uint32(response);
+    buffer << response;
 }
 
 void WorldPackets::GmTicket::GmTicketCreateResponse::AppendBodyTo(ByteBuffer& buffer) const
 {
-    buffer << uint32(response);
+    buffer << response;
 }
 
 void WorldPackets::GmTicket::GmTicketSystemStatus::AppendBodyTo(ByteBuffer& buffer) const
 {
-    buffer << uint32(status);
+    buffer << status;
 }
